refactor(keygen): extract general Sei key derivation out of seitask
The BIP-39 → cosmos secp256k1 → bech32 derivation is general and reused
by the test harness, so move it to a k8s-free internal/keygen package
(Derive() Identity). internal/seitask/keygen keeps only the
seitask-runner's Secret + workflow-vars writer, layered on top via
keygen.Derive — and is the only part the cutover deletes.

hd.go + hd_test.go move verbatim (git mv); Run now calls keygen.Derive.
